Extract common code from benchmarking/nursery/*
